fix: fix checkstyle

+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/repository/jooq/JooqLinkRepository.java
@@ -1,13 +1,11 @@
 package ru.tinkoff.edu.java.scrapper.repository.jooq;
 
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.jooq.DSLContext;
 import ru.tinkoff.edu.java.scrapper.domain.jooq.tables.pojos.ChatLink;
 import ru.tinkoff.edu.java.scrapper.domain.jooq.tables.pojos.Link;
-
-import java.util.List;
-
 import static ru.tinkoff.edu.java.scrapper.domain.jooq.tables.ChatLink.CHAT_LINK;
 import static ru.tinkoff.edu.java.scrapper.domain.jooq.tables.Link.LINK;
 
@@ -17,10 +15,9 @@ public class JooqLinkRepository {
 
     private final DSLContext dslContext;
 
-
     public List<Link> findAllLinksInChat(long chatId) {
         return dslContext.select().from(LINK.join(CHAT_LINK).on(CHAT_LINK.LINK_ID.eq(LINK.ID)))
-                .where(CHAT_LINK.CHAT_ID.eq(chatId)).fetchInto(Link.class);
+            .where(CHAT_LINK.CHAT_ID.eq(chatId)).fetchInto(Link.class);
     }
 
     public void addChatLink(long chatId, Link link) {
@@ -33,13 +30,13 @@ public class JooqLinkRepository {
 
     public List<ChatLink> getChatLinksByUrlAndChatId(long chatId, String url) {
         return dslContext.select().from(CHAT_LINK, LINK).where(LINK.URL.eq(url)
-                .and(LINK.ID.eq(CHAT_LINK.LINK_ID))
-                .and(CHAT_LINK.CHAT_ID.eq(chatId))).fetchInto(ChatLink.class);
+            .and(LINK.ID.eq(CHAT_LINK.LINK_ID))
+            .and(CHAT_LINK.CHAT_ID.eq(chatId))).fetchInto(ChatLink.class);
     }
 
     public List<ChatLink> getChatLinksByLinkId(long linkId) {
         return dslContext.select().from(CHAT_LINK, LINK).where(LINK.ID.eq(linkId)
-                .and(LINK.ID.eq(CHAT_LINK.LINK_ID))).fetchInto(ChatLink.class);
+            .and(LINK.ID.eq(CHAT_LINK.LINK_ID))).fetchInto(ChatLink.class);
     }
 
     public void addLink(String url) {
@@ -47,7 +44,7 @@ public class JooqLinkRepository {
     }
 
     public void removeChatLink(ChatLink link) {
-        dslContext.deleteFrom(CHAT_LINK).where(CHAT_LINK.CHAT_ID.eq(link.getChatId()).and(CHAT_LINK.LINK_ID.eq((long) link.getLinkId()))).execute();
+        dslContext.deleteFrom(CHAT_LINK).where(CHAT_LINK.CHAT_ID.eq(link.getChatId()).and(CHAT_LINK.LINK_ID.eq(link.getLinkId()))).execute();
     }
 
     public void removeLastLink(long linkId) {
